Move extension plugin to app
