+++ packages/cli-kit/src/plugins.test.ts
@@ -1,10 +1,4 @@
-import {
-  getListOfExtensionPoints,
-  getListOfExtensionSpecs,
-  getListOfFunctionSpecs,
-  getListOfTunnelPlugins,
-  runTunnelPlugin,
-} from './plugins.js'
+import {getListOfTunnelPlugins, runTunnelPlugin} from './plugins.js'
 import {err, ok} from './public/common/result.js'
 import {describe, expect, it, vi} from 'vitest'
 import {Config} from '@oclif/core'
@@ -138,101 +132,5 @@ describe('runTunnelPlugin', () => {
     // Then
     expect(got.isErr() && got.error.type).equal('unknown')
     expect(got.isErr() && got.error.message).equal('message')
-  })
-})
-
-describe('getListOfExtensionSpecs', () => {
-  it('returns empty list when there are no extension specs', async () => {
-    // Given
-    const config = new Config({root: ''})
-    vi.spyOn(config, 'runHook').mockResolvedValue({successes: [], errors: []} as any)
-
-    // When
-    const got = await getListOfExtensionSpecs(config)
-
-    // Then
-    expect(got).toEqual([])
-  })
-
-  it('returns list of extension specs', async () => {
-    // Given
-    const config = new Config({root: ''})
-    vi.spyOn(config, 'runHook').mockResolvedValue({
-      successes: [
-        {result: {firstSpec: 1}, plugin: {name: 'extension-spec'}},
-        {result: {secondSpec: 2}, plugin: {name: 'extension-spec-2'}},
-      ],
-      errors: [],
-    } as any)
-
-    // When
-    const got = await getListOfExtensionSpecs(config)
-
-    // Then
-    expect(got).toEqual([{firstSpec: 1}, {secondSpec: 2}])
-  })
-})
-
-describe('getListOfExtensionPoints', () => {
-  it('returns empty list when there are no extension points', async () => {
-    // Given
-    const config = new Config({root: ''})
-    vi.spyOn(config, 'runHook').mockResolvedValue({successes: [], errors: []} as any)
-
-    // When
-    const got = await getListOfExtensionPoints(config)
-
-    // Then
-    expect(got).toEqual([])
-  })
-
-  it('returns list of extension points', async () => {
-    // Given
-    const config = new Config({root: ''})
-    vi.spyOn(config, 'runHook').mockResolvedValue({
-      successes: [
-        {result: {firstPoint: 1}, plugin: {name: 'extension-point'}},
-        {result: {secondPoint: 2}, plugin: {name: 'extension-point-2'}},
-      ],
-      errors: [],
-    } as any)
-
-    // When
-    const got = await getListOfExtensionPoints(config)
-
-    // Then
-    expect(got).toEqual([{firstPoint: 1}, {secondPoint: 2}])
-  })
-})
-
-describe('getListOfFunctionSpecs', () => {
-  it('returns empty list when there are no function specs', async () => {
-    // Given
-    const config = new Config({root: ''})
-    vi.spyOn(config, 'runHook').mockResolvedValue({successes: [], errors: []} as any)
-
-    // When
-    const got = await getListOfFunctionSpecs(config)
-
-    // Then
-    expect(got).toEqual([])
-  })
-
-  it('returns list of function specs', async () => {
-    // Given
-    const config = new Config({root: ''})
-    vi.spyOn(config, 'runHook').mockResolvedValue({
-      successes: [
-        {result: {firstSpec: 1}, plugin: {name: 'function-spec'}},
-        {result: {secondSpec: 2}, plugin: {name: 'function-spec-2'}},
-      ],
-      errors: [],
-    } as any)
-
-    // When
-    const got = await getListOfFunctionSpecs(config)
-
-    // Then
-    expect(got).toEqual([{firstSpec: 1}, {secondSpec: 2}])
   })
 })
